Mal: Practical 4
a) Program to create a file in a directory and  perform following file operations Write into a file Read from a file Delete a file.

b)  Create a new project and create a login Activity. In this create a login UI asking user email and password with an option of remember me checkbox. Also a button displaying Sign In or Register using shared preferences