Mal: Practical 7
A) Create an application that 1) Turn on and off the Bluetooth. 2) Enable and disable Wi-Fi in your phone.

Note: For Bluetooth use API 30 & for wifi use API 27

B) Creates a basic Mobile application that allows you to navigate through the Google maps and also send a push notification and giving a message about updating Google maps