Mal: Practical 8
A) Create a basic application that allows you to download HTML from a given web page using HttpURLConnection.

Note: Use Android 8 Emulator

B) Create an application to parse the data using JSONObject methods and set it in the Text View’s.( employee name and salary stored in JSON format)

C) Write a basic application to (use volley library), create a button and on click of the button a HTTP request will be send to server. The response from the server is then displayed using Toast on the screen.